requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
discord.py==2.3.2
distlib==0.3.8
filelock==3.13.3
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pipenv==2023.12.1
platformdirs==4.2.0
virtualenv==20.25.1
yarl==1.9.4
